Deployment/requirements.txt
streamlit
statsmodels
pandas
scikit-learn
matplotlib
pillow
requests